added employee assignee to each ticket
